chore: Use Dependabot to Upgrade NPM Dependencies (#32)
# Pull Request

<!-- Please use British English for spelling and terminology where
possible. However, if you're more comfortable using another variant
don't worry about it! -->

## Description

This pull request updates the `.github/dependabot.yml` file to improve
the configuration for dependency updates. The most important changes
include modifying the schedule and target branch for GitHub Actions and
adding configuration for npm dependencies.

Configuration updates:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L7-R12):
Modified the schedule to run daily and set the target branch to `main`
for GitHub Actions.
*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L7-R12):
Added configuration for npm dependencies with a daily schedule and
`main` as the target branch.

fixes #39