ci: Add JavaScript CodeQL Analysis (#33)
# Pull Request

<!-- Please use British English for spelling and terminology where
possible. However, if you're more comfortable using another variant
don't worry about it! -->

## Description

This pull request introduces a new GitHub Actions workflow to perform
CodeQL analysis on the repository. The workflow is designed to run on
pushes and pull requests to the `main` branch, as well as on a scheduled
basis.

GitHub Actions workflow addition:

*
[`.github/workflows/codeql-analysis.yml`](diffhunk://#diff-63bd641104d10e25f141d518a16b22a151d125e12701df2f9e79734b23b90188R1-R30):
Added a new workflow named "Code QL Analysis" that runs CodeQL analysis
on `main` branch pushes, pull requests, and on a weekly schedule. It
includes steps to checkout the repository, initialize CodeQL, and
perform the analysis using the `github/codeql-action` actions.

fixes #43