chore: Add Code of Conduct (#119)
# Pull Request

## Description

This pull request includes the addition of a Contributor Covenant Code
of Conduct to the project. The document outlines the community's
commitment to providing a harassment-free experience for everyone and
details the standards and enforcement responsibilities.

Key changes:

* Added a comprehensive Code of Conduct to `docs/CODE_OF_CONDUCT.md` to
ensure a welcoming and inclusive community environment.

The document includes sections on the community's pledge, standards of
behavior, enforcement responsibilities, scope, enforcement guidelines,
and attribution.

Fixes #64