ci: Rename GitHub Actions Steps (#136)
# Pull Request

## Description

This pull request includes changes to the `.github/workflows` files to
improve the clarity of the step names in the GitHub Actions workflows.
The most important change is the renaming of the "Checkout" step to
"Checkout Repository" to provide a more descriptive name.

Changes to GitHub Actions workflows:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L22-R22):
Renamed the "Checkout" step to "Checkout Repository" in three different
jobs.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L22-R22)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L104-R104)
[[3]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L124-R124)
*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L19-R19):
Renamed the "Checkout" step to "Checkout Repository".