ci: Regularly Clear Caches (#139)
# Pull Request

## Description

This pull request introduces a new GitHub Actions workflow to clean
caches on a scheduled basis and manually if needed. The workflow is
defined in the `.github/workflows/clean-caches.yml` file and includes
the following key changes:

GitHub Actions Workflow:

*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dR1-R24):
Added a new workflow named "Clean Caches" that runs on a monthly
schedule and can also be triggered manually. It includes steps to
checkout the repository and delete all caches using the `gh cache
delete` command.