chore: Add CodeQL configuration (#141)
# Pull Request

## Description

This pull request includes changes to the CodeQL configuration and
workflow files to enhance security and quality checks. The most
important changes involve updating the CodeQL configuration to exclude
certain queries and modifying the workflow to use this new configuration
file.

Changes to CodeQL configuration:

*
[`.github/other-configurations/codeql-config.yml`](diffhunk://#diff-687a0af9316072cbdab79cdb046aac181231f8b5f3b5502bc796a9d949ce65beR1-R3):
Added a query filter to exclude the `actions/unpinned-tag` query.

Updates to workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R114):
Configured the CodeQL analysis to use the newly added configuration
file.